pyproject.toml
[project]
name = "assign2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "camelot-py>=1.0.9",
    "fastapi>=0.128.0",
    "fastembed>=0.7.4",
    "ipykernel>=7.1.0",
    "langchain>=1.2.3",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.1.0",
    "pgvector>=0.3.6",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
